Update migration method with feedback from @larry0x.
