+++ Cargo.toml
@@ -8,7 +8,7 @@ license       = "Apache-2.0"
 repository    = "https://github.com/CosmWasm/cw-nfts"
 homepage      = "https://cosmwasm.com"
 documentation = "https://docs.cosmwasm.com"
-rust-version = "1.65"
+rust-version  = "1.65"
 
 [workspace.dependencies]
 cosmwasm-schema = "1.1.5"
