+++ contracts/cw721-base/README.md
@@ -16,7 +16,7 @@ Implements:
 The `ExecuteMsg` and `QueryMsg` implementations follow the [CW721 spec](../../packages/cw721/README.md) and are described there.
 Beyond that, we make a few additions:
 
-* `InstantiateMsg` takes name and symbol (for metadata), as well as a **Minter** address. This is a special address that has full 
+* `InstantiateMsg` takes name and symbol (for metadata), as well as a **Minter** address. This is a special address that has full
 power to mint new NFTs (but not modify existing ones)
 * `ExecuteMsg::Mint{token_id, owner, token_uri}` - creates a new token with given owner and (optional) metadata. It can only be called by
 the Minter set in `instantiate`.
@@ -30,44 +30,11 @@ or just create a custom contract as the owner and use that contract to Mint.
 
 If provided, it is expected that the _token_uri_ points to a JSON file following the [ERC721 Metadata JSON Schema](https://eips.ethereum.org/EIPS/eip-721).
 
-## Updating the minter
-
-For contract versions &\gt 0.16$ the minter has been replaced by an
-owner which is updatable via the two-step ownership transfer process
-of [cw_ownable](https://crates.io/crates/cw-ownable). To retreive the
-owner, `QueryMsg::Ownership {}` may be executed.
-
-`QueryMsg::Minter {}` has not bee removed, though after version 0.16
-the response type has made the minter field optional as it may be
-unset. For all intents and purposes, whenever the word minter is used,
-it means owner, and owner in turn means minter, $minter \iff owner$.
-
-Before 0.16:
-
-```rust
-pub struct MinterResponse {
-    pub minter: String,
-}
-```
-
-After 0.16:
-
-```rust
-pub struct MinterResponse {
-    pub minter: Option<String>,
-}
-```
-
-NFTs on version 0.16 may upgrade via `MigrateMsg::From016 {}`. For an
-example of doing so, see
-[this](https://github.com/CosmWasm/cw-nfts/blob/zeke/updatable-minter/contracts/cw721-base/src/multi_tests.rs#L83)
-integration test.
-
 ## Running this contract
 
-You will need Rust 1.44.1+ with `wasm32-unknown-unknown` target installed.
+You will need Rust 1.65+ with `wasm32-unknown-unknown` target installed.
 
-You can run unit tests on this via: 
+You can run unit tests on this via:
 
 `cargo test`
 
@@ -80,7 +47,7 @@ ls -l cw721_base.wasm
 sha256sum cw721_base.wasm
 ```
 
-Or for a production-ready (optimized) build, run a build command in the 
+Or for a production-ready (optimized) build, run a build command in the
 repository root: https://github.com/CosmWasm/cw-plus#compiling.
 
 ## Importing this contract
@@ -89,7 +56,7 @@ You can also import much of the logic of this contract to build another
 CW721-compliant contract, such as tradable names, crypto kitties,
 or tokenized real estate.
 
-Basically, you just need to write your handle function and import 
+Basically, you just need to write your handle function and import
 `cw721_base::contract::handle_transfer`, etc and dispatch to them.
 This allows you to use custom `ExecuteMsg` and `QueryMsg` with your additional
 calls, but then use the underlying implementation for the standard cw721
