+++ contracts/cw721-base/examples/schema.rs
@@ -1,13 +1,12 @@
 use cosmwasm_schema::write_api;
 use cosmwasm_std::Empty;
 
-use cw721_base::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
+use cw721_base::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 fn main() {
     write_api! {
         instantiate: InstantiateMsg,
         execute: ExecuteMsg<Empty, Empty>,
         query: QueryMsg<Empty>,
-        migrate: MigrateMsg,
     }
 }
