+++ contracts/cw721-base/src/upgrades/mod.rs
@@ -1,0 +1,1 @@
+pub mod v0_16;
